blueprints to labelers (deprecation)
