+++ bot/handlers/invite_help.py
@@ -1,14 +1,14 @@
-from vkbottle.bot import Blueprint, Message
+from vkbottle.bot import BotLabeler, Message
 
 from config import GROUP_ID
 
-bp = Blueprint("Invite event")
-bp.labeler.vbml_ignore_case = True
+bl = BotLabeler()
+bl.vbml_ignore_case = True
 
 HELP_LINK = "https://github.com/timius100/HuTao-Bot#команды"
 
 
-@bp.on.message(action="chat_invite_user")
+@bl.message(action="chat_invite_user")
 async def invite_event_reaction(message: Message):
     """React when bot is invited to a chat"""
     if message.action.member_id == -GROUP_ID:
@@ -21,7 +21,7 @@ async def invite_event_reaction(message: Message):
         )
 
 
-@bp.on.message(text="!помощь")
+@bl.message(text="!помощь")
 async def help_handler(message: Message):
     await message.answer(
         "Основные команыды:\n"
@@ -34,7 +34,7 @@ async def help_handler(message: Message):
     )
 
 
-@bp.on.private_message(text="Начать")
+@bl.private_message(text="Начать")
 async def start_private_handler(message: Message):
     """
     This handler only works in private messages
