+++ bot/handlers/ping.py
@@ -1,12 +1,12 @@
 import time
 
-from vkbottle.bot import Blueprint, Message
+from vkbottle.bot import BotLabeler, Message, rules
 
-bp = Blueprint("Test ping")
-bp.labeler.vbml_ignore_case = True
+bl = BotLabeler()
+bl.vbml_ignore_case = True
 
 
-@bp.on.message(text=("!пинг", "!гнш пинг"))
+@bl.message(text=("!пинг", "!гнш пинг"))
 async def do_ping(message: Message):
     command_start = time.time()
     command_end = time.time()-command_start
